feat(schemas): add UnicodeTreeNode and UnicodeTree schemas
* Introduced `UnicodeTreeNodeSchema` to represent file and directory nodes.
* Added `UnicodeTreeSchema` for a tree structure of files and directories.
* Removed unused `schemas.ts` file to streamline the codebase.
* Updated imports in `openapi.ts` and `index.ts` to reflect the new schema structure.
