+++ packages/schemas/src/index.ts
@@ -4,5 +4,5 @@ export type { ApiError } from "./api";
 export { FileEntryListSchema, FileEntrySchema, UCDStoreManifestSchema } from "./fs";
 export type { FileEntry, FileEntryList, UCDStoreManifest } from "./fs";
 
-export { UnicodeVersionListSchema, UnicodeVersionSchema } from "./unicode";
-export type { UnicodeVersion, UnicodeVersionList } from "./unicode";
+export { UnicodeTreeNodeSchema, UnicodeTreeSchema, UnicodeVersionListSchema, UnicodeVersionSchema } from "./unicode";
+export type { UnicodeTree, UnicodeTreeNode, UnicodeVersion, UnicodeVersionList } from "./unicode";
